requirement.txt
genshin
python-dotenv
